// ww_frontend/src/app/Models/OutletAndPromoDTO.ts
export interface OutletAndPromoDTO{

    storeName: string
    unitNumber: string
    phoneNumber:string
    postalCode: string
    outletPicUrl:string
    numberOfPromos:string
    itemName: string
    description: string
    startDate: Date
    endDate: Date
    startTime: Date
    endTime: Date
    status: string
    price: string
    promoPicUrl:string
    promotionId: string
    addressId: string
    lat:number
    lng:number
    searchedAddLat:number
    searchedAddLng: number

}